requirements.txt
google-generativeai>=0.8.0
requests>=2.25.1
python-dotenv>=0.19.0
rich>=14.0.0
typing-extensions>=4.0.0
flask>=2.3.0
flask-socketio>=5.3.0
